refactor(trade_ideas): extract read-model contracts from the service
service.py (1258 lines) defined the trade-idea read models (TradeIdeaView,
TradeIdeaQueryPage, TradeIdeaListQuery/Result/SortKey) and the not-found /
duplicate / pre-approval errors alongside the audited lifecycle service. Move
the inert contracts into trade_ideas/service_models.py (1258->1188) and
re-export.

Only inert data types moved; all audited service logic stays in
TradeIdeaService. Re-export keeps identity stable across service, service_models,
and the package __init__ (verified), so the CLI/TUI adapters are unaffected. 274
trade_ideas tests pass.
